+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added auto page height for multiple pages (for `section` or after custom page break)
 - Added type validation for parameters in method `createPdf`
+- Added support `SVGElement` object for `svg` node (`SVGElement` object is available only in browser)
 - Fixed a bug in the write method where it did not wait for the file write operation to complete
 - Fixed SVG loading
 - Fixed rendering SVG without viewBox